[mxfp8 moe training] move generic dispatch helper from mxfp8 -> tensor.py
stack-info: PR: https://github.com/pytorch/ao/pull/3853, branch: danielvegamyhre/stack/139
