+++ torchao/prototype/moe_training/__init__.py
@@ -1,7 +1,9 @@
-from torchao.prototype.moe_training.mxfp8_grouped_mm import (
-    _quantize_then_scaled_grouped_mm,
+from torchao.prototype.moe_training.fp8_grouped_mm import (
     _to_fp8_rowwise_then_scaled_grouped_mm,
     _to_mxfp8_then_scaled_grouped_mm,
+)
+from torchao.prototype.moe_training.tensor import (
+    _quantize_then_scaled_grouped_mm,
 )
 
 __all__ = [
